refactor: split preview_ingress.rs god-file into modules
